+++ lib/valid_forms.dart
@@ -5,6 +5,6 @@
 library valid_forms;
 
 export './src/valid_form.dart';
-export './src/valid_form_field.dart';
-export './src/confirmable_valid_form_field.dart';
-export './src/valid_field_validator.dart';
+export './src/valid_input.dart';
+export './src/confirmable_valid_input.dart';
+export './src/valid_input_validator.dart';