+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.3.0
+
+### Change
+
+- `ValidFormField --> ValidInput`.
+- `ValidTextFormField --> ValidTextInput`.
+- `ConfirmableValidFormField --> ConfirmableValidInput`.
+- `ValidFieldValidator --> ValidInputValidator`.
+
 ## 0.2.0
 
 ### Add